backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
typedb-driver
opensearch-py
openai
pandas
openpyxl
pdfplumber
langchain
langchain-text-splitters
requests
websockets
